+++ src/open/render_map_cells.c
@@ -28,10 +28,10 @@ void render_map_cell(s32 col, s32 row, KfCellVisibility cell)
         object_index += KF_MAP_MESHES_PER_BANK;
     }
     setVector(&position,
-        col * KF_MAP_TILE_SIZE - (u16)open_graphics_runtime.render_state.view_position.vx,
+        col * KF_MAP_TILE_SIZE - open_graphics_runtime.render_state.view_position.vx,
         map_floor_height_grid.cells[row][col] * -KF_MAP_HEIGHT_STEP -
-            (u16)open_graphics_runtime.render_state.view_position.vy,
-        row * KF_MAP_TILE_SIZE - (u16)open_graphics_runtime.render_state.view_position.vz);
+            open_graphics_runtime.render_state.view_position.vy,
+        row * KF_MAP_TILE_SIZE - open_graphics_runtime.render_state.view_position.vz);
     if (orientation == KF_ENUM_ENCODE(u8, KF_MAP_ORIENT_QUARTER_TURN) - 1) {
         position.vz += KF_MAP_TILE_SIZE;
     } else if (orientation == KF_ENUM_ENCODE(u8, KF_MAP_ORIENT_HALF_TURN) - 1) {
